requirements.txt
alabaster==0.7.10
appdirs==1.4.0
appnope==0.1.0
argh==0.26.2
asgi-redis==1.2.0
asgiref==1.0.1
asn1crypto==0.22.0
attrs==16.3.0
autobahn==0.17.2
Automat==0.5.0
Babel==2.4.0
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2017.4.17
cffi==1.9.1
channels==1.1.3
chardet==3.0.4
click==6.7
CommonMark==0.5.4
constantly==15.1.0
coreapi==2.3.0
coreapi-cli==1.0.6
coreschema==0.0.4
cryptography==1.8.1
cycler==0.10.0
daphne==1.1.0
decorator==4.0.11
Django==1.10.5
django-appconf==1.0.2
django-braces==1.11.0
django-extensions==1.7.9
django-googledrive-storage==1.3.4
django-haystack==2.6.1
django-http-proxy==0.4.3
django-jsonfield-compat==0.4.4
django-oauth-toolkit==0.12.0
django-polymorphic==1.1
django-rest-framework-social-oauth2==1.0.5
django-rest-swagger==2.1.2
django-sphinxdoc==1.5.1
django-webpack-loader==0.4.1
djangorestframework==3.6.3
djangorestframework-jwt==1.10.0
djangorestframework-yaml==1.0.3
djoser==0.5.4
docutils==0.13.1
drfdocs==0.0.11
drfpasswordless==1.0.8
earthengine-api==0.1.114
enum34==1.1.6
epydoc==3.0.1
epyparse==0.2.5
ERAlchemy==1.1.0
functools32==3.2.3.post2
future==0.16.0
GDAL==2.1.0
geojson==1.3.5
google-api-python-client==1.6.2
gunicorn==19.6.0
guzzle-sphinx-theme==0.7.11
httplib2==0.10.3
idna==2.5
imagesize==0.7.1
incremental==16.10.1
ipaddress==1.0.18
ipdb==0.10.2
ipython==5.3.0
ipython-genutils==0.1.0
itypes==1.1.0
Jinja2==2.9.6
jsonfield==2.0.1
jsonpickle==0.9.4
jsonschema==2.6.0
livereload==2.5.1
MarkupSafe==1.0
matplotlib==2.0.0
meld3==1.0.2
msgpack-python==0.4.8
numpy==1.12.1
oauth2client==4.0.0
oauthlib==2.0.1
openapi-codec==1.3.1
packaging==16.8
pandas==0.20.1
pathlib2==2.2.1
pathtools==0.1.2
pexpect==4.2.1
pickleshare==0.7.4
plantweb==1.0.1
pockets==0.5.1
port-for==0.3.1
prompt-toolkit==1.0.13
ptyprocess==0.5.1
pyasn1==0.2.2
pyasn1-modules==0.0.8
pycparser==2.17
pycrypto==2.6.1
# Comment until the upgrade to 2.1.3
# Use GDAL as alternative
# Install locally for upload to GAE
# pip install -t lib/ pygdal==2.1.2.3
# pygdal==2.1.2.3
Pygments==2.2.0
pygraphviz==1.3.1
PyJWT==1.5.0
pyOpenSSL==16.2.0
pyparsing==2.1.10
python-dateutil==2.6.0
python-openid==2.2.5
pytz==2017.2
PyYAML==3.12
recommonmark==0.4.0
redis==2.10.5
requests==2.18.1
requests-file==1.4.2
requests-oauthlib==0.8.0
retrying==1.3.3
rsa==3.4.2
scandir==1.5
service-identity==16.0.0
simplegeneric==0.8.1
simplejson==3.10.0
singledispatch==3.4.0.3
six==1.10.0
snowballstemmer==1.2.1
social-auth-app-django==1.1.0
social-auth-core==1.2.0
Sphinx==1.6.2
sphinx-autoapi==0.4.0
sphinx-autobuild==0.6.0
sphinxcontrib-dd==0.1.7
sphinxcontrib-dotnetdomain==0.4
sphinxcontrib-golangdomain==0.1.3.dev0
sphinxcontrib-mermaid==0.2.1
sphinxcontrib-napoleon==0.6.1
sphinxcontrib-plantuml==0.8.1
sphinxcontrib-swaggerdoc==0.1.4
sphinxcontrib-websupport==1.0.1
SQLAlchemy==1.1.11
subprocess32==3.2.7
supervisor==3.3.1
tornado==4.5.1
traitlets==4.3.2
Twisted==17.1.0
txaio==2.6.1
typing==3.6.1
Unidecode==0.4.20
uritemplate==3.0.0
urllib3==1.21.1
watchdog==0.8.3
wcwidth==0.1.7
whitenoise==3.3.0
Whoosh==2.7.4
zope.interface==4.3.3
